requirements.txt
alembic==1.13.1
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
backports.tarfile==1.0.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
build==1.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
docopt==0.6.2
docutils==0.21.1
executing==2.0.1
fastapi==0.110.1
fastjsonschema==2.19.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
ipython==8.12.3
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.0
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keyring==25.1.0
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.2.0
mysql==0.0.3
mysqlclient==2.2.4
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nh3==0.2.17
packaging==24.0
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pipdeptree==2.16.2
pipreqs==0.5.0
pkginfo==1.10.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pydantic==1.10.15
Pygments==2.17.2
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-multipart==0.0.9
python-socketio==5.11.2
pywin32==306
pywin32-ctypes==0.2.2
pyzmq==25.1.2
readme_renderer==43.0
redis==5.0.3
referencing==0.34.0
reflex==0.4.7
reflex-hosting-cli==0.1.11
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
rpds-py==0.18.0
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
sqlmodel==0.0.16
stack-data==0.6.3
starlette==0.37.2
starlette-admin==0.13.2
tabulate==0.9.0
tinycss2==1.2.1
tomlkit==0.12.4
tornado==6.4
traitlets==5.14.2
twine==5.0.0
typer==0.12.3
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.20.0
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
yarg==0.1.9
zipp==3.18.1
